requirements.txt
Flask>=1.1.2
python-nmap>=0.6.4